wrong file count fixed

+++ mobile-app/lib/backend/storage/storage_repository.dart
@@ -3,6 +3,8 @@ import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:amplify_storage_s3/amplify_storage_s3.dart';
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/services.dart';
+import 'package:mobile_app/backend/Blocs/sync/sync_bloc.dart';
+import 'package:mobile_app/backend/Blocs/sync/sync_events.dart';
 import 'package:mobile_app/backend/repositories/LocalDataRepository.dart';
 import 'package:mobile_app/utils/connectivity.dart';
 import 'package:path_provider/path_provider.dart';
@@ -10,6 +12,34 @@ import 'package:path_provider/path_provider.dart';
 import 'dataStorePaths.dart';
 
 class StorageRepository {
+  static SyncBloc? syncBloc;
+
+  static Future<bool> hasConnectivity() async {
+    if (_last_connectivity_check != null) {
+      if (_last_connectivity_check!.difference(DateTime.now()).inSeconds < 2) {
+        if (_had_connectivity != null) {
+          return Future.value(_had_connectivity);
+        }
+      }
+    }
+
+    try {
+      final result = await InternetAddress.lookup('google.com');
+      if (result.isNotEmpty && result[0].rawAddress.isNotEmpty) {
+        _had_connectivity = true;
+      } else {
+        _had_connectivity = false;
+      }
+    } on SocketException catch (_) {
+      _had_connectivity = false;
+    }
+    _last_connectivity_check = DateTime.now();
+    return _had_connectivity!;
+  }
+
+  static bool? _had_connectivity;
+  static DateTime? _last_connectivity_check;
+
   static Future<void> downloadFile(File toDownload, String path,
       {bool checkConnection = true, DateTime? lastModifiedOnline}) async {
     try {
@@ -23,8 +53,17 @@ class StorageRepository {
         }
       }
 
+      syncBloc?.add(StartLoadingFileEvent());
+
+      //check internet connection
+      if (!await hasConnectivity()) {
+        syncBloc?.add(CancelSyncEvent());
+        return;
+      }
+
       await Amplify.Storage.downloadFile(key: path, local: toDownload);
       //set local last changed to onlyine last changed
+      syncBloc?.add(LoadedFileEvent());
 
       if (lastModifiedOnline != null) {
         await toDownload.setLastModified(lastModifiedOnline);
@@ -44,6 +83,8 @@ class StorageRepository {
 
       print("file for $path successfully downloaded");
     } catch (e) {
+      syncBloc?.add(LoadedFileEvent());
+      //todo: handle with sync bloc
       print("not found $path");
       print(e.toString());
     }
@@ -61,13 +102,27 @@ class StorageRepository {
           return dataStorePath;
         }
       }
+
+      //add to sync bloc
+      syncBloc?.add(StartLoadingFileEvent());
+
+      if (!await hasConnectivity()) {
+        syncBloc?.add(CancelSyncEvent());
+        return dataStorePath;
+      }
+
       final result = await Amplify.Storage.uploadFile(
         local: file,
         key: dataStorePath,
       );
 
+      //remove from sync bloc
+      syncBloc?.add(LoadedFileEvent());
+
       return result.key;
     } catch (e) {
+      syncBloc?.add(LoadedFileEvent());
+      //remove from sync bloc
       return dataStorePath;
     }
   }
@@ -85,12 +140,26 @@ class StorageRepository {
     try {
       // RemoveOptions options =
       // RemoveOptions(accessLevel: StorageAccessLevel.guest);
+
+      //send to sync bloc
+      syncBloc?.add(StartLoadingFileEvent());
+
+      if (!await hasConnectivity()) {
+        syncBloc?.add(CancelSyncEvent());
+        return path;
+      }
+
       final result = await Amplify.Storage.remove(
         key: path,
         // options: options
       );
+
+      syncBloc?.add(LoadedFileEvent());
+
       return result.key;
     } catch (e) {
+      //remove from sync bloc
+      syncBloc?.add(LoadedFileEvent());
       rethrow;
     }
   }
@@ -136,6 +205,8 @@ class StorageRepository {
       await localCacheFile.delete();
       return true;
     } catch (e) {
+      print('error in db save operation');
+      print(e.toString());
       return false;
     }
   }
